backend/requirements.txt
setuptools
flask
Flask_SQLAlchemy==2.4.0
flask_restful
flask_cors
ethereum
mysqlclient
hexbytes
web3
dateutils
